backend/requirements.txt
# pip3 install spacy
# python3 -m spacy download en_core_web_sm

fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
# pydantic==2.7.4
pydantic==2.10.6

numpy==1.26.4 
pandas==2.2.2